Deleted boundary dependecies from plug-in View.
Now the View doesn't depend on the List<> boundary but to
draw the living objects used an Iterator.

It has also been added a "canDrawLine" flag in the DrawModel class
that allows the renderer to understand if he can draw the line or not.

In this version has been removed the points LinkedList<> from the
TrajectoryCalculator class.