Successfully handle bubble's direction.
The current version allow the user to throw a Bubble and see it
moves in the right direction.

To do so two classed have been added:
- LinearBoundary
- Direction

The first, defines a Boundary object that allow the MotionRuler
to detetched if the bubble hit the boundary.

The second one defines a single direction along a precise angle.
The DirectionHandler will check the right direction of the moving
object and get it from the Direction object.